+++ server/Content/docs/ReleaseNotes.md
@@ -12,9 +12,8 @@ Bugs
 * Case sensitivity / insensitivity issues accross the board, esp. related to JSON serialization of package names.
 * Default user ID / authentication shouldn't be in the NapackSettings.json file, but somewhere else reasonably common and accessible.
 * Significantly increase unit testing and functional testing, so that the website can be tested locally with the in-memory store.
-* The napack icon needs professional attention.
 * Search is missing the 'click for details' function on retrieved items linking to API, versions, authors, etc.
-* Requests don't log IPs or do any attempt at throttling.
+* Request IP throttling internals and country geolocation database reading work.
 
 Core Functionality
 ------------------
@@ -25,6 +24,7 @@ Core Functionality
 * The Napack Server Framework needs a webpage for Napack upload/update and user registration. Update [Creation.md](./Creation.md) and [Registration.md](Registration.md) when done.
 * Backend email setup on Napack to support email validation; fix the default email and settings at that time.
 * Create a tree-view page to navigate through Napacks and their dependencies.
+* Create a system stats page listing requests per country for the current day.
 
 Final Release Steps
 -------------------